+++ CHANGELOG.md
@@ -6,6 +6,7 @@ _2024-XX-XX_
 
 - Location names are now displayed in the minimap when hovering over a section on the schedule.
 - Custom sections can now have custom start and end dates, which affects the events when exporting your schedule to calendar.
+- View a unit breakdown on your plan by clicking on the total units number at the bottom of the plan.
 - The school to which a course or section belongs is now displayed in the course or section information menu. This is available for courses and sections offered since 2024 Fall.
 - Copy document ID to the clipboard when clicking on the document name in the toolbar.
 - Copy share link to the clipboard when clicking on the people icon in the toolbar, which is visible when a document is persistently shared.